fix: split drm ports
